isLogEmpty V2
Summary:
The current isLogEmpty implementation is increasingly complex, and yet sensitve
to too many nodes being unhealthy or down. A better way to get this information
is if we can have the sequencer maintain it. As the sequencer now periodically
gets the trim point, this diff enables clients to directly ask the sequencer
whether a log is empty. If the sequencer is not done activating, or is in some other
error state, it will return E::AGAIN. The interval for periodic polling of trim points can
be changed by adjusting --get-trimpoint-interval server setting.

Also updates `ld-is-log-empty` script to have a `--use-v2` option, as well as a
`--version-result-comparison`.

Reviewed By: runemaster

Differential Revision: D14656365

fbshipit-source-id: 90536f679aa8c3a2ee28e64c412fd290ac5aae76