Remove REQUIRES_RESTART from the max-random-delay-shard-is-rebuilt-message setting
Summary: Reading through the code, I can't see a reason why this flag requires a restart. If it got changed, new scheduled SHARD_IS_REBUILT messages will use the new setting. If you want to apply it on already scheduled SHARD_IS_REBUILT that's when you need a restart. So in general, I think it's ok to remove this flag and let the users decide what they want to do.

Reviewed By: shobhitdayal

Differential Revision: D15984785

fbshipit-source-id: fcda9ee180d7770ff94ddd8d242dc571671e3069