Explicitly shut the processor down before returning from NodesConfigurationInit
Summary: If something in one of the requests (e.g. WorkerCallbackHelper.h) extends the lifetime of the processor, it may lead to its destruction on one its workers causing a crash. Let's explicitly drain the processor first to make sure that all the requests are done and new ones are not added.

Reviewed By: davidvigier, shixiao

Differential Revision: D16180237

fbshipit-source-id: e13cb6e90986d3c94a97474e2b77291abb030901