Add another stat for rocksdb blocks read by iterators
Summary:
Right now we export:
 - number of *bytes* read from disk,
 - number of *blocks* taken from block cache.

These two are not comparable (without knowing block size, which is not easy). We'd probably prefer to export both stats in bytes, but rocksdb doesn't expose cache hits in bytes. So this diff adds a third counter:
 - number of *blocks* read from disk.

This allows roughly estimating cache hit rate by dividing block counts. Then we can assume that the hit ratio in bytes is about the same as in block count (not a safe assumption either!) and estimate how much disk IO the block cache is saving.

Reviewed By: scsiguy

Differential Revision: D14692558

fbshipit-source-id: 5619ef7dc1221042ecaf56c47f193d2773e35b4f