Fix Start onreceived to skip processing message if socket was closed while waiting.
Summary:
Permission checks is a async call and connection can be closed while that is
happening. Start onreceived accesses ClientID and once Socket is closed it
should not be used.

In the new model, onReceived will be processed asynchronously and hence socket
can be closed while the message waits in the queue. Fixing this problem by
trying to get socket proxy before initiating any type of processing of START
message inorder to skip processing if it is found that socket was closed.

Other option was to cahnge registerOnClosedCallback to schedule to invoke the
callback immediately. I checked the code base and found most of the places
except these two handle the registerOnClosedCallback. Also none of the other
conditions call this on receiving a message they call it as part of send part.

Reviewed By: davidvigier

Differential Revision: D15487136

fbshipit-source-id: d7608c77e3bf3ced565aa87e066bb9b30ceea70a