Fix registerOnClosed to check if the socket was closed and isClosed API to Sender
Summary:
There is a hole where a onclose callback can be registered even after socket
was closed. Because the socket is immediately removed from the maps. Fixing
this by checking if the socket was closed before adding the callback.

Added API to sender to check if socket to given address isClosed. Used that api from START_onreceived to detect closed sockets.

Reviewed By: davidvigier

Differential Revision: D16295229

fbshipit-source-id: 0d70442adfa1e1b28b574fa58d0dc08e23e4bf70