NodesConfiguration: Allow config with empty nodes section to pass the validation
Summary:
In the new cluster membership world, the nodes configuration will be the source of truth of the nodes in the cluster.

Optimally, when nodes configuration manager is enabled and is the source of truth, we shouldn't be parsing the whole node section at all, but it's a bit tricky because to know the settings you need to finish parsing the config. We can do it the same way the logsconfig is currently doing it, but it's a bit hacky and requires separating the NodesConfig into its own updateable. So for now, let's just allow configs with empty nodes section to pass the validation.

Reviewed By: Zepheus

Differential Revision: D16377038

fbshipit-source-id: a774ed89bc28ebf07c70f432ca37e8ebb39b453d