Fixed a bug, updated tests, added some more stats.
Summary:
When the log gets enqueued for evaluation for sequencer reactivation, an additional flag is enabled in LogState if it was queued by the timer callback. So that it is not postponed more than once. The bug was that it was not being switched off after the processing.

Tested this out on a real cluster. Also updated the tests and added some more stats.

Reviewed By: runemaster

Differential Revision: D16343038

fbshipit-source-id: 39b4cb5f3b797b10c240186844ae414f0a3bc98b