Fix GSS to not send check-seals if sequencer is PREEMPTED by epoch store
Summary:
GSS sends check-seals to detect preemption even if sequencer is already in PREEMPTED state. This is done since we are not sure if the 'preempted_by' is actually the most recent preemptor.
However, if sequencer doesn't have nodeset information, it won't make any progress indefinitely, in this case it's better to just send a redirect to whatever node epoch store replied with.

Reviewed By: runemaster

Differential Revision: D16270910

fbshipit-source-id: 9191b4938b08ed51f442467b39374c8d5bdc2e4b