ConnectThrottle to not depend on socket or settings
Summary:
ConnectThrottle's simple logic used to depend on socket to get setting. Socket
depends on connect throttle. Breaking the circular dependency.

With this we lose the ability to update values to existing objects on the fly
which I felt was not necessary anyways.

This is part of my plan to separate out Socket and connection into it's
separate target so that we could plugin different socket implementations
easily.

Reviewed By: mcrnic

Differential Revision: D16213480

fbshipit-source-id: d99a7061319973a83782eefc051bbbf1ce20fadf