Allow registering socket close callback on closed server socket.
Summary:
Same instance of closed server socket will be used to re initiate connection to
remote. Hence, even a socket closed callback be registered it should be ok and
it is guaranteed to be invoked.

In case of incoming sockets the socket instance is deleted once the socket is
closed hence disallowing registering socket callback on them.

Updated comments.

Reviewed By: davidvigier

Differential Revision: D16428796

fbshipit-source-id: a3c30581ddbcbac80cbf5d59c87d3ea23942fab0