ClientReadersFlowTracer: don't report lag when on tail
Summary: Readers on tail might improperly be marked as lagging on ClientReadersFlowTracer when the Sequencer respond with an incorrect timestamp for the tail record (which happens for example when all records are trimmed). This diff forces the tracer to always record a lag of 0 when we are at tail to prevent this situation.

Reviewed By: tau0

Differential Revision: D16486998

fbshipit-source-id: 4c93a7b5ff238c490775938830e11c75428afed4