Use `LocationScope' enum in `getReplicationInfo' [1/2]
Summary:
For some obscure historical reasons we used to return tolerable failure domain as a plain string whereas the rest of our APIs use `LocationScope` enum.

This stack of changesets makes `TolerableFailureDomain` struct use the`LocationScope` enum for the domain representation and also makes the `narrowest_replication` field use that enum as a key.

**This changeset**
1. Rename old string fields: `narrowest_replication` -> `narrowest_replication_legacy`, `TolerableFailureDomain.domain` -> `TolerableFailureDomain.domain_legacy`.
2. Add `narrowest_replication` and `TolerableFailureDomain.domain` fields that use `LocationScope`.

Reviewed By: MohamedBassem

Differential Revision: D16440741

fbshipit-source-id: de25113dd08c42da062aa9df208adb31e4641d9d