NodesConfiguration: Adding a new PROVISIONING storage state
Summary:
Nodes need to ack the health of their local storage before getting enabled. This diff introduces a new storage state PROVISIONING that all the nodes start into. Once the shard writes RebuildingCompleteMetadata to its local storage, it should apply the MARK_SHARD_PROVISIONED transition to transition to NONE. Shards can never go back to this state.

Shards that are int the PROVISIONING state won't get enabled by the Maintenance manager. Also shards will use this state to skip rebuilding on startup if they are in this state (instead of relying on generation == 1).

The changes introduced by this diff:
- A new initial state PROVISIONING is introduced.
- I renamed PROVISION_SHARD to BOOTSTRAP_ENABLE_SHARD to better indicate its new functionality and to avoid confusion with the PROVISIONING state.
- BOOTSTRAP_ENABLE_SHARD: Will change nodes from NONE to RW, instead of adding nodes.
- MARK_SHARD_PROVISIONED: is the new transition that will move shards from PROVISIONING to NONE.
- ADD_EMPTY_SHARD: Will add the nodes in the new PROVISIONING state.
- I removed the EMPTY_ VERSION constraint from BOOTSTRAP_ENABLE_SHARD (in the next diff I'll add the bootstrapping flag that will restrict the usage of this transition).
- I changed the semantics of BOOTSTRAP_ENABLE_SHARD to respect if the node was added as a metadata node.

Reviewed By: runemaster

Differential Revision: D16418234

fbshipit-source-id: e2f9064c84913467dbfda2d6df0d17a8b0c0a7c3