NodesConfiguration: Removing the hasAllUpdates restriction on the initial NC update
Summary: Removing the hasAllUpdates restriction on the initial NC update. This restriction doesn't hold in the self registration world. The first node will add itself but won't update at least the metadata nodes section, the admin server will probably be the one doing so. So this restriction doesn't really apply in here and it's not adding extra benefits because the config will still be valid.

Reviewed By: runemaster

Differential Revision: D16335515

fbshipit-source-id: 01a5a12d19438ad56d3880b495c06d8f5bd14ce7