log NC in NCPublisher and log line when diverging with same version
Summary:
This diff logs the published NC in the `nc_update` field; logs to stderr when
NCM NC and ServerConfig NC diverge with the same version. It also changes the
`is_same` calculation during logging a sample to logger to ignore
`last_change_timestamp`, since the NCM NC could be locally constructed or the
ServerConfig NC could stamp it with the time when the NC gets converted from
the legacy NodesConfig format, thus having a different timestamp.

Reviewed By: runemaster

Differential Revision: D16628491

fbshipit-source-id: 5c9f476b9ef216cbb8aae66cc3232252c02709cb