EASY / FUN: explicitly specify node id in addNewNodeUpdate
Summary:
Currently we use a static variable to auto-increment the node id in tests. This
is brittle because when different subset of the tests are run or when the order
of the tests change, the behavior of the tests change accordingly.

This diff forces all callsites to explicitly pass in the node id.

Reviewed By: runemaster

Differential Revision: D16570786

fbshipit-source-id: 17e6796fb6633443b01dac7413c27ae178c0c6a5