Admin command table to compile socket as separate target
Summary:
Convert to print peer_name in InfoSocketTable depended on
Sender::describeConnection API. As we are making Socket independent of Sender
that API and Socket target also will be independent of Sender hence that API is
not available.

Adding describeConnection to SocketDeps (done already) Here using that api to
describe connection.

Reviewed By: mcrnic

Differential Revision: D16274690

fbshipit-source-id: e3af75dd8b752f385365d1578d66ed1bd068d9f8