Fix logsdb iterator pinning memtable for too long
Summary:
When seeking above highest LSN, PartitionedRocksDBStore::Iterator used to leave its data_iterator_ untouched, which violated a rule for iterators used on read path: seek() must either refresh or destroy pre-existing subiterators - otherwise the subiterators can pin memtables for longer than iterator-cache-ttl.

I introduced the bug in D15350079; it was intended as an optimization (not central to that diff, just along the way).

This diff just undoes most of the optimization: if logsdb iterator is seeked to above the last record, we'll just destroy the data iterator in current partition. If afterwards more records are written and iterator is seeked again - too bad, we'll have to create a new data iterator in same partition. That's how it used to be before D15350079, and we were fine. Another option would be to try harder and only destroy data_iterator_ if it hasn't been seeked for more than iterator-cache-ttl, but doesn't seem worth the effort; such effort is better spent on T45309029, which is the proper solution to this class of problems.

Reviewed By: scsiguy

Differential Revision: D16728028

fbshipit-source-id: 57a72fd0bb72dd322f764544a23056f489625249