Sending FIFO order replies from sequencer for stream appends.
Summary:
To provide FIFO ordering guarantee, the reply from sequencer for a stream append must happen only after all previous ESNs have been reliably stored on the nodes. For normal appends ACK to the client can happen in any order.

In this diff we are modifying the logic to send reply back in `Appender::onReaped` for write stream appends. By this time all previous `Appender` belonging to the same epoch should have been reaped. This directly provides FIFO ordering for write stream appends belonging to the same epoch.

Reply to a client must happen on the same worker where `APPEND_Message` was received. So, we post it to the corresponding worker if `Appender::onReaped` is being invoked on a different worker which can happen in the current design.

Reviewed By: al13n321

Differential Revision: D16676908

fbshipit-source-id: f4287b92ab91d4be098f2308081fc75f8e75a508