Fix read-stream-guaranteed-delivery-efficiency getting stuck when a node is down but not in rebuilding set
Summary: It used to stall until all fully authoritative nodes chime in. So if a node is down, the streams get stuck until it's added to rebuilding set and marked unrecoverable. This diff makes the stream proceed with an f-majority, as long as the remaining shards are in SCD filtered out list.

Reviewed By: scsiguy

Differential Revision: D16824372

fbshipit-source-id: 5d991c10ee9f483674a5d4e7ac0c905708318480