Fix ClientReadStream getting stuck when a node is down and another node has underreplicated ranges
Summary: If a node reports underreplicated range, we don't immediately add it to SCD filtered out list right away, in hopes that the node is not actually missing any relevant records (because underreplicated ranges are usually way over-reported). The blacklisting+rewind are only done if we all nodes chimed in, and we still have a gap. But the definition of "all nodes" wasn't quite right: without this diff, we only blacklist+rewind if all fully authoritative nodes chimed in; if one node is down, and another node has underreplicated range and is missing a record we need - we'll get stuck instead of rewinding. This diff fixes it: now we blacklist+rewind if we heard from all nodes that are not blacklisted; this is the same condition as for rewinding to all-send-all.

Reviewed By: scsiguy

Differential Revision: D16824369

fbshipit-source-id: 694af2dec3b892da01da7680b5804fe9a43e68c7