New target for connection
Summary:
Create new target for connection and socket this will help us in replacing
choosing between asyncsocket and libevent socket implementation. Some more work
is needed there. But this is a start.

What I don't like is that SocketDependencies dependency is half baked in this
target. Will slowly work on resolving that as part of SocketDependencies refactor but I have no ideas at this point.

Reviewed By: mcrnic

Differential Revision: D16274734

fbshipit-source-id: 486a5c8ab95687b42580fa217f3cd4e33b8450cf