Modifying ACK behavior of stream write append sink
Summary: Currently, `StreamWriterAppendSink` can ACK back to the `BufferedWriter` in any order it wants and stream resume bit is set for a message that is one more than the max ACKed seq num. We change this to max prefix ACKed seq num as follows. When we do get an ACK for a seq num, we check if it is `max_prefix_acked_seq_num_ + 1`, and if not we store the record and status and exit. Else we callback immediately and trigger as many callbacks that are ready to be called.

Reviewed By: al13n321

Differential Revision: D16837093

fbshipit-source-id: 861725e6e0047a643e669610fb03fe29562d6b04