ClientReadersFlowTracer: fix issues with time lag correction in lagging readers metric
Summary:
tl;dr This change makes all past bucket lag corrections be "consolidated"
before moving on the next bucket and fixes a flipped sign.

In current behavior, we place a positive value in the bucket where a client
started rejecting records, and a negative value when client started accepting
record again in hopes that total sum will include both values. For example, if
a reader has time lag "10" at time/bucket 1 (when client was rejecting records)
and time lag "20" at time/bucket 5 (when client started accepting records
again), we add 10 to time correction at bucket 1 and subtract 20 at time bucket
5, so when we sum from 1 to 5, the final summation will include the -10 lag
correction for the lag that was accumulated in between those instants.

However if time bucket 1 is evicted, we will miss the addition of +10 and the
value will be incorrect.

This change makes it so that old buckets always have valid time lag corrections
while we maintain an incomplete value (with only a positive correction) in the
newest bucket.

Reviewed By: al13n321

Differential Revision: D16584195

fbshipit-source-id: 2354d9e0c0e5bf7933ded6a99659340e423c457d