Assigning write stream id
Summary: As a placeholder, we were using log id as the same as stream id. In this diff, we generate a random 64 bit value for write stream id using `folly::Random::rand64()`. Based on analysis the probability that two clients clash on a write stream id for the same log is very very low.

Reviewed By: al13n321

Differential Revision: D17066970

fbshipit-source-id: f46732870541d8ad8c43c6c3047ea6a413aeb14e