Garbage collection for write streams in epoch sequencer
Summary:
Adding GC to EpochSequencer so that we can bound the additional memory used to support write streams. Memory limit can be set in `Settings` and will be directly passed on to the `EpochSequencer` via `EpochSequencerImmutableOptions`.

We have at most 2 `EpochSequencer` objects lying around in the `Sequencer` and so, an appropriate bound set for `EpochSequencer` will also limit the total memory budget used to support streams.

Also, replaced per-stream mutex with a global per-epoch mutex. This is not as bad as it seems. Primarily because the additional cost is just that we synchronize LSN allotment among different concurrent streams that belong to the same log. Note that this still does not impact non-stream workloads. On the positive side, we save close to 48 bytes per stream by eliminating a per-stream mutex and hence memory will not be a bottleneck anymore to support many more write streams.

Reviewed By: runemaster

Differential Revision: D17116825

fbshipit-source-id: b55dbf9d2018dbd90c474d7edaf5e8ccf38aeda7