fix the bias selection due to ignoring the domain storage state in WeightedCopysetSelector
Summary:
D15981223 the behavior of WeightedCopySetSelector has changed
so that we ignore storage state in calculating domain
weights. However, this leads to biased distribution behavior in case
that a domain has all nodes disabled (e.g., in RO) except for one.
This patch addressed the problem by properly taking account of
storagte state in calculating domain weights.

This patch also makes clarification on the definition of storage
capacity in storage config, which is a per-shard attribute. A method
of getWritableStorageCapacity() is added to calcualte the sum of all
writeable capacity of all shards in a node.

Reviewed By: al13n321

Differential Revision: D17177532

fbshipit-source-id: 8820aa67e3665f2a79df5a52c89e457849351973