A high but sane default for deletetion rate-limit
Summary:
Picking a high default for this settings. Some of the flash clusters are still missing it, although they may not be susceptible to this problem.

I chose a value that is way higher than the write-throughput on any shard on either flash or disk. Did I get it right?

Reviewed By: al13n321

Differential Revision: D17185764

fbshipit-source-id: 02627b22a9a19a0f642f7975c732626ac1deb8c2