Pass shared_ptr for sequencers list to SequencerLocator instead of raw pointer
Summary: Use shared_ptr instead of raw pointer to pass the sequencers config to the SequencerLocator from SequencerRouter to avoid dangling pointers if the SequencerRouter happens to be destroyed while fetching the log attributes. The sequencers config is copied before blacklisting more sequencers if needed, to avoid modifying the objects that may be accessed from a different thread.

Reviewed By: MohamedBassem

Differential Revision: D17214153

fbshipit-source-id: 398b377a7f8a9f362292c7463a12dd6f47aab195