Authenticate servers during handshake if configured
Summary:
Added an optional "cluster_node_identity" attribute in security_information block of the config to specify the identity that servers should present to clients in their SSL certificate. This is used by client to authenticate servers.
Added server authentication in ACK_Message::onReceived. The handler checks that the SSL certificate has the expected identity and fails with E::ACCESS otherwise.

Reviewed By: MohamedBassem

Differential Revision: D17207608

fbshipit-source-id: 6480ec3b79223e115f09a6bb692371641e140219