Constrain most logsdb directory iterators to a single log
Summary:
For performance and simplificatoin. Should prevent T52832400-like performance issue for hot path directory reads.

Before this diff, when reading logsdb directory of a single log we would use an iterator that sees the whole metadata column family, so after moving the iterator you'd need to check that it's (a) valid, (b) on a directory entry, (c) the log ID in the directory entry matches our log ID. This diff makes use of rocksdb's iterate_lower_bound and iterate_upper_bound options to make iterator only see a range of keys corresponding to logsdb directory of the log we're looking for. Benefits:
 - This makes the code a little simpler because now you only need to check (a).
 - Makes data consistency checks a little stronger because any malformed directory entries in the key range are unexpected and will raise an alarm, whereas previously they were mostly treated as case (b) and ignored.
 - Mostly avoids a rocksdb performance landmine when lots of consecutive keys are deleted: rocksdb iterator has to do a linear scan through the deleted keys to skip them. If all logsdb directory entries were deleted, most iterator operations used to do this scan through directory of all logs. Now it'll scan just one log - still suboptimal, but ~100x improvement.

We didn't do it this way originally (a thousand years ago) because iterate_lower_bound and prefix seek didn't exist, and iirc iterate_upper_bound support was incomplete.

Reviewed By: scsiguy

Differential Revision: D17128401

fbshipit-source-id: 4cd362f1e15604207e3caeffe76015e1f285f4e8