Don't crash on InfoConfig command when the config doesn't have a source
Summary:
If the ServerConfig originates from the client through config synchronization or from the node itself before it knew about its NodeID, `getServerOrigin()` will be an invalid NodeID and calling `index()` on that triggers an ld_check. For the purpose of the hotfix, let's just return -1 if we don't know the source, and in a later diff, I'll just kill the whole server origin thing because it's useless.

Fixes #106

Reviewed By: Zepheus

Differential Revision: D17318908

fbshipit-source-id: 0ea2cd2f5ce8549e3cde704ebb95c3db77c60641