Create OverloadDetector
Summary:
This diff introduces OverloadDetector, a client-side overload detector that is
used to filter out reader health metrics from clients that are overloaded.

Every period (1min by default):

1/ We dispatch sampling "requests" randomly spread across that period, that
retrive the occupancy of recv-q in sockets;
2/ We check the occupancy of the xth percentile (by default x=99) of recently
used sockets to assess if we are overloaded.

A socket is considered "recently used" if the number of bytes it received is at
least y times the size of its window (by default, y=1.0) or if its occupancy is
above threshold (when we don't expect it to be drained any time soon).

Reviewed By: mcrnic

Differential Revision: D17184777

fbshipit-source-id: 358616886e3da931c6ab9cadfc97eb7b98a23dcc