CircleCI: Don't update modules from upstream and use the commited hashes instead
Summary:
`--remote` updates submodules from upstream. This used to be ok because we used to be building against master which is almost always in sync with the remote thanks to shipit. On the other hand, this broke the release branch because some breaking changes landed in upstream.

This diff stops fetching from upstream in circle CI and uses the committed hashes in the repo instead.

Reviewed By: AhmedSoliman

Differential Revision: D17365456

fbshipit-source-id: 332c24ca25d0f4f6fcc029b2265ec46b0af145b4