Add CheckpointStore class with update functionality.
Summary:
The first version of CheckpointStore. It's using VersionedConfigStore, where the key is the pair customer_id, log id and the value is its LSN. Currently, the CheckpointStore provides only updating LSNs. I haven't implemented yet dealing with versions (just passing folly::none as an argument to VersionedConfigStore).

TODO(T53779383): Currently, the key in the VersionedConfigStore is serialized pair (customerID, logID) and the value is LSN.
We want the key to be the customerID and the value to be a map (logID -> LSN). That would be done in a separate diff.

Reviewed By: MohamedBassem

Differential Revision: D17282619

fbshipit-source-id: 54f70285d4472eae59228abd4b4b16bf970e8823