Introduce folly::IOBuf in PayloadHolder.
Summary:
In an effort to move to libevent1.0 we need to make minimal use of evbuffer and
bufferevent in core data path. IOBuf are performant and ultimately AsyncSocket
will be providing us with IOBuf after readin from socket.

For both these reasons, we need PayloadHolder to stop using evbuffer and switch
to IOBuf.

No zero copy anymore:
Earlier we were calling evbuffer ZeroCopied buffers and we had ways to dispose
of those evbuffer on right thread. This was a requirement for evbuffer and also
a good idea, but it's not necessary now after introduction of IOBuf. Btw, the
evbuffer anyways was not zero copy as a pullup was always done which could
lead to copy if data size was more than 1K.
Now, buffer is always copied from evbuffer into IOBuf. This IOBuf is moved to
whichever thread needs it and deleted once last PayloadHolder reference is
dropped.

Reviewed By: mcrnic

Differential Revision: D15802109

fbshipit-source-id: 393fd2606024fafa15ee71115783d389f097b2a3