Use IOBuf on write path when serializing messages.
Summary:
Write serialized output IOBuf instead of evBuffer. This IOBuf will be moved into AsyncSocket in later diffs.

This diff introduces a small regression for libevent2 based sockets, as the buffer is copied from IOBuf to evbuffer. I can optimize that by saving the IOBuf in Envelope and destroying it as part of onBytesPassed if you think that is necessary.

Reviewed By: mcrnic

Differential Revision: D15808112

fbshipit-source-id: 0a03d25ead23b8f1b1a0544299033100be5156cd