Add support for TLS sensing in ConnectionListener
Summary:
TLS connections are handled by instantiating 2 listeners on different ports - one for plain and other for encrypted data. This is because we explicitly specify to the ConnectionListener which type is it. DATA or DATA_SSL.

In order to unify these ports, we need to support a listener detecting if connection starts with a TLS handshake.

ConnectionListener.cpp creates a new event after accept is called. This event either timeouts, or a callback is triggered by a read event on the desired file descriptor. This callback peeks the first 3 bytes and detects TLS handshakes 1.1-1.3. This information is then passed on wards via NewConnectionRequest.

Reviewed By: mcrnic

Differential Revision: D17260530

fbshipit-source-id: c39872a9f84dabccb518e51d4743404ce73d2980