Fix trimming, enable in tupperware
Summary:
The implementation of `shouldTrim()` always returned false. I've fixed that to match what we do in the other RSMs. Also, enabling snapshotting in tupperware spec.

I also set the default value for keeping extra deltas before the snapshot point to zero, this means that we will trim everything before the snapshot point, not sure why it was set to 14 days. This is how LogsConfigManager is also configured. No point of keeping extra records there.

Reviewed By: MohamedBassem

Differential Revision: D17419615

fbshipit-source-id: ffcc83bcfcee1ec2588c59d09a0c7fc69b500f40