Stop reporting boycotting controller state on boycotts
Summary:
Boycotting used to report its in memory state when a boycotting decision is taken. This was a complicated process, as it needs to communicate back with the Stats controller to get this info. Furthermore, this info is also outdated because that's the in memory current stats of the controller, and is not necessarily the one that we considered this node outlier based on.
Nobody looks at this stats anyways, as you can infer them from the dashboards, so let's just kill this complicated approach.

Reviewed By: tau0

Differential Revision: D17480041

fbshipit-source-id: 41feba9ab296293af89a22b69772edc41a0464ad