Make EvTimer compatible with AsyncTimeout interface
Summary:
Socket class has various timers which use libevent structures and use the api through SocketDependencies. But as we are transitioning to AsyncSocket we will need the socket timers which need to work with libevent2 and libevent1 apis.

Idea is to convert all these timers to EvTimer and wrap libevent2 and libevent1 api inside the EvTimer class. The implementation will be switched according to a setting.

EvTimer will have folly::AsyncTimeout interface, this will automatically convert EvTimer to folly::AsyncTimeout in libevent1 version.
EvTimerLegacy is implementation of EvTimer interface with libevent2. Its almost similar to libeventtimer.
EvTimerWithFolly inherits from AsyncTimeout and is basically AsyncTimeout.

This diff also helps in transistioning LibeventTimer to libevent1 transparently. A later diff will also convert timer within LibeventTimer to EvTimer. As a result, LibeventTimer does not need to move to HHWheelTimer immediately thus preserving performance.

Reviewed By: mcrnic

Differential Revision: D16733291

fbshipit-source-id: 252e472d7cbb7f8fa0d1edcf2810771acb727667