Fix flakiness of the remove node test
Summary:
The test was flaky because although the node is disabled in the config, it doesn't have a maintenance which allowed maintenance manager to enable it back. The remove call was then failing because the node was in NONE_TO_RO instead of NONE.
This diff, applies a maintenance on the node to prevent this scenario from happening.

Reviewed By: AhmedSoliman

Differential Revision: D17453326

fbshipit-source-id: 1340a0a3a3210cee222d052baa1417b97f9bb905