Treat NONE_TO_RO storage state as ShardOperationalState::ENABLING
Summary: As in description. getMaintenanceProgressInternal depends on getShardOperationalStateInternal to determine if maintenance has completed by checking the target state against the current shard operational state. If a node was being enabled and then before it could transition to RO (current state in NC: NONE_TO_RO), if a disabled maintenance was added and MM hasn't reacted to this, querying ShardOperationalState will return DRAINED which is the target state for the new disable maintenance which in turn will incorrectly return that maintenance completed in the call to getMaintenanceProgressInternal. Fix is to treat NONE_TO_RO as ShardOperationalState::ENABLING

Reviewed By: MohamedBassem

Differential Revision: D17492951

fbshipit-source-id: a61f32e9da8e1ab07a155c4a824e7dec40a753f4