Fix UnreleasedRecordDetectorTest.TransientSequencerFailure flakiness
Summary:
It was expecting to get CONNFAILED but got NOSEQUENCER. I guess what happens is AppendRequest fails to send to one sequencer node (with CONNFAILED), blacklists that node, retries sequencer locator, and sequencer locator says NOSEQUENCER because the only sequencer node is blacklisted. I'm not sure why this doesn't make NOSEQUENCER always override other return codes (except TIMEDOUT), not sure what the intended behavior is, not sure how this test used to work before (assuming it ever worked), and not sure what each of the million error codes returned from AppendRequest/Client::append() are intended to mean.

This diff just works around it by removing the weird retry loop and replacing it with a more targeted wait for listener initialization.

Reviewed By: davidvigier

Differential Revision: D17534062

fbshipit-source-id: d73c375d2858d4572104b831bb1fb9cf334de079