Filter out nodes rebuiling in RELOCATE mode to remove bias and increase speed of drains
Summary:
Cf https://fb.workplace.com/groups/LogDeviceOncall/permalink/2557610250954241/

I'm adding a new filtering logic where shards in RELOCATE mode are filtered out
by default by SCD so the record can be picked up by some one else. This will
help reduce a bias that causes a huge imbalance when we drain a failure domain.

If doing so means all shards in the copyset are considered down by the filtering
logic, we resort to the old mechanism. In other words, if a node being drained
has the last remaining copy, it will be the donor.

This diff also removes the logic by which a recipient ships the copy even if it
sees itself as down, which is logic we only want during reads but not during
rebuilding as it conflicts with the change discussed above.

I'd like to have feedback on that last point as I saw a comment explaining that
the logic that I'm removing would sometimes trigger if there is a combo of shard
with dirty ranges and shard in RESTORE mode, and this causes over-replication.
To me it sounds fine to remove the logic but I would like Justin's opinion.

It's unfortunate that the logic is duplicated both for rebuilding v1 and v2. I
initially wanted to refactor the code in D15876283 but this ended up harder than
I thought and I did not go further than just extracting the logic out of
LogRebuilding which is already bloated.

I am going for an approach that is simple for now (just try to filter the
relocate shards and fallback to not doing that if necessary). T47692209
describes ideas for making the selection more fair.

The "--filter-relocate-shards" setting causes logdeviced to include a
FILTER_RELOCATE_SHARDS flag in the SHARD_NEEDS_REBUILD record. This flag
instructs all donor nodes that for this rebuilding version, we use the new
filtering logic. Note that for drains, we will have to modify the FBJE handlers
as well as MM to make sure they include the flag. This is tracked in T54390368.

== Rollout notes ==

* Deploy this package everywhere in prod;
* Ensure all client binaries are re-deployed so they understand the new format;
* Enable --filter-relocate-shards so that rebuilding triggered by the server follow the new behavior;
* Modify MM and fbje to ensure the FILTER_RELOCATE_SHARDS flag is used for drains.

== Rollback notes ==

Ensure at least logdevice version n-1 has this diff before enabling the feature;

It is safe to disable the feature while rebuilding is ongoing. The current
rebuilding will continue with the new behavior. The next rebuilding will not
observe the new behavior.

Reviewed By: al13n321

Differential Revision: D15877267

fbshipit-source-id: 6b7f9f774eea4f790ebc6bce1dd42f129f6734f0