Introduce EvBase, Event and EvTimer class.
Summary:
Switching between libevent2 and libevent1 using a setting will be good to allow rolling back quickly if changes did not work in prod. This gives us great insurance policy. It helps in comparing performance quickly without doing an extra deploy across cluster.

I believed earlier that it was not ok to have libevent2 and libevent1 version in the same build. But that was not true I could have libevent2 and libevent1 in the same binary because of EvBase. EventLoop depends on EvBase which can be libevent2 based EvBaseLegacy or libevent1 based EvBaseWithFolly.

What we just need to make sure is if EventLoop is created with a EvBaseLegacy then all the Event and EvTimer on this EventLoop are also the Legacy version. Similarly if a EventLoop is created with EvBaseWithFolly then Event and EvTimers created on this EventLoop are folly or libevent1 version.

This diff introduces EvBase which wraps libevent1 and libevent2 event_bases. It also introduces Event and EvTimer that uses EvBase type to instantiate the right version of Event or EvTimer.

Other thing it helped me as I was implementing folly::AsyncSocket support is that it made me independent of Nick's diffs as I could instantiate listeners with Legacy EventLoop implementation and thus allow me run all the integration tests. But this benefit is not long term.

Reviewed By: mcrnic

Differential Revision: D17063068

fbshipit-source-id: 6f78a839f37d09fca6aabdbd96e4d27b9f45bc81