Fix a bug in AdminCommandClient when connection fails inline
Summary: If the socket connection fails inline (e.g. when the unix socket doesn't exist), we will get stuck in the event_base::loopOnce call until the command timeout kicks in which causes tests to timeout. This diff checks first if there are any connections that are still pending before waiting on the event loop.

Differential Revision: D17600881

fbshipit-source-id: ebb4d4ddb9cbfc624825e41001d44136f561ab9e