Fix a bug in AdminCommandClient when connection fails inline (#114)
Summary:
Pull Request resolved: https://github.com/facebookincubator/LogDevice/pull/114

If the socket connection fails inline (e.g. when the unix socket doesn't exist), we will get stuck in the event_base::loopOnce call until the command timeout kicks in which causes tests to timeout. This diff checks first if there are any connections that are still pending before waiting on the event loop.

Reviewed By: tau0

Differential Revision: D17600881

fbshipit-source-id: 6778a440654ef4467cc9245ba5b7e3dd97083739