Make Appender not send waves too quickly if stores are failing quickly, and fix AppendIntegrationTest.GraylistCorruptedStorageNodes flakiness
Summary:
Without this diff, Appender::onRecipientFailed() sends a new wave immediately. This is a bit scary: if some storage nodes keep quickly failing STOREs but keep getting selected into copysets (e.g. if there's an f-majority of such nodes), Appender will send waves as fast as it can. This can probably overload the cluster and maybe get into a positive feedback loop where all these extra waves are themselves causing the STORE failures. This diff makes it send new wave immediately only if current wave is <= 2, otherwise wait for store timeout.

Also fixes the integration test that was failing because gray-list-threshold was too small, and because one append wasn't enough to ensure that N0 is in graylist. It also was hitting the quick retry scenario from previous paragraph, although idk if that would actually cause the test to fail after the other two issues are fixed.

Reviewed By: mcrnic

Differential Revision: D17585513

fbshipit-source-id: f118429a800b2122183ae415282d67676f2cb24e